Convert helper files to TS
